--- src/views/components/AudioCallApp/IAudioCallList.ts
@@ -1,6 +1,0 @@
-import { IWord } from '../../../helpers/interfaces';
-
-export interface IAudioCallList extends IWord {
-  testAnswerList: string[];
-  userAnswer: boolean;
-}
