--- src/views/components/Navbar/INavbar.ts
@@ -1,6 +1,0 @@
-export interface INavbar {
-  render(url: string | null): Promise<string>;
-  after_render(): Promise<void>;
-  setLoginState: () => void;
-  setLogoutState: () => void;
-}
